\name{clValid-internal}
\alias{stability}
\alias{vClusters}
\alias{connectivity}
\alias{dunn}
\alias{BHI}
\alias{BSI}
\alias{matchGO}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal clValid functions}
\description{
  Internal clValid functions
}
\details{
  These functions are not to be called directly by the user
}

\keyword{internal}